Refactor: broke functions down into separate files for easier readability.
