--- src/components/PlayerButtons.test.js
@@ -1,8 +1,0 @@
-const player_hand = require("./PlayerButtons")
-
-test("Properly sums hand after drawing aces", () => {
-    expect
-
-
-
-});
